it_fusion_modular_project/MARIA-master/requirements.txt
omegaconf==2.3.0
numpy==1.26.0
pandas==2.2.2
tqdm==4.64.0
scipy==1.13.1
pillow==10.4.0
missingpy==0.2.0
xgboost==1.7.1
pandarallel==1.6.1
scikit-learn==1.3.0
hydra-core==1.3.0
pynrrd==0.4.3
category-encoders==2.6.1
imbalanced-learn==0.12.2
scikit-survival==0.23.0
